cvforgebot/requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
jinja2>=3.0.0
aioredis>=2.0.0
python-latex>=1.0.0
redis>=5.0.0 